Some server reorganization and cleanup.
