+++ lib/api.js
@@ -22,13 +22,25 @@ const restaurants = mongodb({
   }
 });
 
-export default feathers()
-  .configure(feathers.rest())
-  .configure(feathers.socketio())
-  .use(bodyParser.json())
-  .use(bodyParser.urlencoded({ extended: true }))
-  .use('/restaurants', restaurants)
-  .use('/orders', mongodb({
-    collection: 'orders',
-    connectionString: config.mongodb
-  }));
+export default function() {
+  let api = feathers()
+    .configure(feathers.rest())
+    .configure(feathers.socketio())
+    .use(bodyParser.json())
+    .use(bodyParser.urlencoded({ extended: true }))
+    .use('/restaurants', restaurants)
+    .use('/orders', mongodb({
+      collection: 'orders',
+      connectionString: config.mongodb
+    }));
+
+  // Override listen because we need the server to set up the API
+  let oldListen = this.listen;
+  this.listen = function() {
+    let server = oldListen.apply(this, arguments);
+    api.setup(server);
+    return server;
+  };
+
+  this.use('/api', api);
+}
