+++ lib/app.js
@@ -1,12 +1,13 @@
 import path from 'path';
 import feathers from 'feathers';
-import config from './config';
 import Steal from 'steal';
-import styles from './styles';
-import inlineCache from './inline_cache';
 import url from 'url';
+
+import config from './config';
 import api from './api';
-import ajax from './ajax';
+import styles from './iso/styles';
+import inlineCache from './iso/inline_cache';
+import ajax from './iso/ajax';
 
 const steal = Steal.clone();
 global.System = steal.System;
@@ -17,7 +18,7 @@ const appPromise = steal.import("app/main.stache!can/view/autorender/system", "a
   .then(function([main, layout]){
 
   return feathers()
-    .use('/api', api)
+    .configure(api)
     .use(ajax())
     .use(feathers.static(path.join(__dirname, '..', 'public')))
     .use("/", function(req, res){
