+++ lib/iso/inline_cache.js
